refactor(ui): split oversized components
