+++ src/ui/components/MarkdownRenderer/useMarkdownInteractions.ts
@@ -1,0 +1,185 @@
+import {
+  type MouseEvent as ReactMouseEvent,
+  type RefObject,
+  useCallback,
+  useState,
+} from "react";
+import { findQuoteOccurrences } from "../../../markup";
+import type { CommentAnchorDraft } from "../../../types/criticmarkup";
+
+export interface RangeCommentDraft {
+  blockIndex: number;
+  top: number;
+  anchor: CommentAnchorDraft;
+}
+
+interface HoveredBlock {
+  index: number;
+  top: number;
+}
+
+function offsetWithinBlock(root: HTMLElement, node: Node, offset: number) {
+  const range = document.createRange();
+  range.selectNodeContents(root);
+  range.setEnd(node, offset);
+  return range.toString().length;
+}
+
+function captureRangeCommentDraft(
+  selection: Selection,
+): RangeCommentDraft | null {
+  if (selection.isCollapsed || selection.rangeCount === 0) {
+    return null;
+  }
+  const range = selection.getRangeAt(0);
+  const startElement =
+    range.startContainer instanceof HTMLElement
+      ? range.startContainer
+      : range.startContainer.parentElement;
+  const block = startElement?.closest<HTMLElement>("[data-block-index]");
+  if (!block) {
+    return null;
+  }
+  const anchorRoot =
+    block.querySelector<HTMLElement>("[data-anchor-root]") ?? block;
+  if (!anchorRoot.contains(range.startContainer)) {
+    return null;
+  }
+  let start = offsetWithinBlock(
+    anchorRoot,
+    range.startContainer,
+    range.startOffset,
+  );
+  let end = anchorRoot.contains(range.endContainer)
+    ? offsetWithinBlock(anchorRoot, range.endContainer, range.endOffset)
+    : (anchorRoot.textContent?.length ?? 0);
+  const plainText = anchorRoot.textContent ?? "";
+  while (start < end && /\s/.test(plainText[start])) {
+    start += 1;
+  }
+  while (end > start && /\s/.test(plainText[end - 1])) {
+    end -= 1;
+  }
+  const quote = plainText.slice(start, end);
+  if (quote.length < 3 || quote.length > 300) {
+    return null;
+  }
+  const occurrences = findQuoteOccurrences(plainText, quote);
+  const occurrence = Math.max(occurrences.indexOf(start) + 1, 1);
+  const blockIndex = Number(block.dataset.blockIndex);
+  if (!Number.isInteger(blockIndex)) {
+    return null;
+  }
+  return {
+    blockIndex,
+    top: block.offsetTop,
+    anchor: { quote, occurrence, start, end },
+  };
+}
+
+export function useMarkdownInteractions(input: {
+  bodyRef: RefObject<HTMLDivElement | null>;
+  canComment: boolean;
+}) {
+  const [hoveredBlock, setHoveredBlock] = useState<HoveredBlock | null>(null);
+  const [rangeCommentDraft, setRangeCommentDraft] =
+    useState<RangeCommentDraft | null>(null);
+
+  const handleBodyMouseOver = useCallback(
+    (event: ReactMouseEvent<HTMLDivElement>) => {
+      if (!input.canComment) {
+        return;
+      }
+      const target = event.target;
+      if (!(target instanceof HTMLElement)) {
+        return;
+      }
+      const block = target.closest("[data-block-index]");
+      if (!(block instanceof HTMLElement)) {
+        return;
+      }
+      const index = Number(block.getAttribute("data-block-index"));
+      const top = block.offsetTop;
+      setHoveredBlock((currentBlock) => {
+        if (
+          currentBlock &&
+          currentBlock.index === index &&
+          currentBlock.top === top
+        ) {
+          return currentBlock;
+        }
+        return { index, top };
+      });
+    },
+    [input.canComment],
+  );
+
+  const handleBodyClick = useCallback(
+    (event: ReactMouseEvent<HTMLDivElement>) => {
+      const target = event.target;
+      if (!(target instanceof Element)) {
+        return;
+      }
+      const footnoteLink = target.closest(
+        "[data-footnote-ref], [data-footnote-backref]",
+      );
+      if (!(footnoteLink instanceof HTMLAnchorElement)) {
+        return;
+      }
+      event.preventDefault();
+      const hash = footnoteLink.getAttribute("href");
+      if (!hash || !hash.startsWith("#")) {
+        return;
+      }
+      document
+        .getElementById(hash.slice(1))
+        ?.scrollIntoView({ behavior: "smooth", block: "center" });
+    },
+    [],
+  );
+
+  const handleBodyMouseUp = useCallback(() => {
+    if (!input.canComment) {
+      return;
+    }
+    const selection = window.getSelection();
+    if (!selection) {
+      return;
+    }
+    const draft = captureRangeCommentDraft(selection);
+    if (draft) {
+      setRangeCommentDraft(draft);
+    }
+  }, [input.canComment]);
+
+  const handleSpecialBlockAnchor = useCallback(
+    (blockIndex: number, anchor: CommentAnchorDraft) => {
+      const block = input.bodyRef.current?.querySelector<HTMLElement>(
+        `[data-block-index="${blockIndex}"]`,
+      );
+      if (!block) {
+        return;
+      }
+      setRangeCommentDraft({ blockIndex, top: block.offsetTop, anchor });
+    },
+    [input.bodyRef],
+  );
+
+  const dismissRangeComment = useCallback(() => {
+    setRangeCommentDraft(null);
+    window.getSelection()?.removeAllRanges();
+  }, []);
+
+  const handleBodyMouseLeave = useCallback(() => setHoveredBlock(null), []);
+
+  return {
+    dismissRangeComment,
+    handleBodyClick,
+    handleBodyMouseLeave,
+    handleBodyMouseOver,
+    handleBodyMouseUp,
+    handleSpecialBlockAnchor,
+    hoveredBlock,
+    rangeCommentDraft,
+  };
+}
